more unsigned long casts

+++ pdld/src/bytearray/bytearray.c
@@ -42,13 +42,13 @@ void bytearray_read_3_bytes (unsigned long *value_p, const unsigned char *src, i
         int i;
 
         for (i = 0; i < 3; i++) {
-            *value_p |= src[i] << (CHAR_BIT * i);
+            *value_p |= (unsigned long)src[i] << (CHAR_BIT * i);
         }
     } else {
         int i;
 
         for (i = 0; i < 3; i++) {
-            *value_p |= src[3 - 1 - i] << (CHAR_BIT * i);
+            *value_p |= (unsigned long)src[3 - 1 - i] << (CHAR_BIT * i);
         }
     }
 }
@@ -67,7 +67,7 @@ void bytearray_read_4_bytes (unsigned long *value_p, const unsigned char *src, i
         int i;
 
         for (i = 0; i < 4; i++) {
-            *value_p |= src[4 - 1 - i] << (CHAR_BIT * i);
+            *value_p |= (unsigned long)src[4 - 1 - i] << (CHAR_BIT * i);
         }
     }
 }
